Added initial support of modifiers. Now anyone can put macro for method and if it has a code of block, it will put `self.name_of_modifier();` call there. Force building of external trait inheritance in tests.
