+++ contracts/access/ownable/impls.rs
@@ -26,14 +26,11 @@ pub trait Ownable: OwnableStorage {
     }
 
     fn renounce_ownership(&mut self) {
-        self.only_owner();
-
         // TODO: Emit event
         *self._owner_mut() = ZERO_ADDRESS.into();
     }
 
     fn transfer_ownership(&mut self, new_owner: AccountId) {
-        self.only_owner();
         assert_ne!(new_owner, ZERO_ADDRESS.into(), "{}", OwnableError::NewOwnerIsZero.as_ref());
         // TODO: Emit event
         *self._owner_mut() = new_owner;
